feat(lora): MLX-native LoKr loader + routing (sc-2216/sc-2314)
Completes the mflux-side LoKr load path so LoKr adapters run natively on MLX
(no torch fallback). LoKrLoader detects networkType=lokr from safetensors
metadata, groups lokr_w1/w2[/_a/_b] by module path, reconstructs the Kronecker
delta (alpha/rank from metadata), and swaps each targeted Linear for a
LoKrLinear. LoRALoader routes per-file: lokr → LoKrLoader, else the existing
A/B pattern flow (unchanged). LoKrLinear casts its delta to the activation
dtype so it composes with bf16/quantized bases.

Key layout verified empirically against PEFT 0.19.1 + write_lokr_adapter: raw
keys `<module>.lokr_w1/w2`, no transformer./base_model.model. prefix, alpha in
metadata. 9 tests green (reconstruction parity + loader mechanics: module
replacement, scale=0 no-op, mx.load metadata round-trip); existing 27 LoRA
tests unaffected.

NOT YET: SceneWorks must drop reject_lokr_loras + the LoKr MLX-dispatch
exclusion and repin to this branch; real trained-adapter validation owed.

+++ src/mflux/models/common/lora/layer/lokr_linear_layer.py
@@ -73,5 +73,7 @@ class LoKrLinear(nn.Module):
 
     def __call__(self, x):
         base_out = self.linear(x)
-        lokr_out = mx.matmul(x, self.delta.T)
+        # The reconstructed delta is held at the factors' precision (typically fp32);
+        # cast to the activation dtype so it composes with a bf16/quantized base.
+        lokr_out = mx.matmul(x, self.delta.astype(x.dtype).T)
         return base_out + self.scale * lokr_out
